enhance: Define plugin with customized printer and parser
