+++ src/main.ts
@@ -1,26 +1,6 @@
 import { format } from 'prettier';
-import { scrapRawScript } from '~/helpers/parser';
-import { parsers as babelParsers } from 'prettier/parser-babel';
-import { parsers as typescriptParsers } from 'prettier/parser-typescript';
-
-import { printers } from './helpers/printer';
-
-export const parsers: { [parserName: string]: any } = {
-  babel: {
-    ...babelParsers.babel,
-    astFormat: 'babel-ast',
-  },
-  typescript: {
-    ...typescriptParsers.typescript,
-    astFormat: 'typescript-ast',
-  },
-};
-
-const myCustomPlugin = {
-  parsers,
-  printers,
-  astFormat: 'estree',
-};
+import { scrapRawScript } from '~/helpers/reader';
+import plugin from './plugin';
 
 export default (async () => {
   const args = process.argv.slice(2);
@@ -29,7 +9,7 @@ export default (async () => {
     const rawScript = await scrapRawScript(args[1]);
 
     const formattedScript = format(rawScript, {
-      plugins: [myCustomPlugin],
+      plugins: [plugin],
     });
     console.log(formattedScript);
   }
